Files were organized in appropriate folders: commands, utilities, models
