+++ shellfoundry/models/shell_template.py
